+++ frontend/src/component/playlist-gallery/playlist.gallery.tsx
@@ -1,39 +1,33 @@
-import React, {useCallback, useState, useRef, useEffect} from 'react';
-import './playlist-tree.scss';
+import ServerConfig from "../../model/server-config";
 import {PlaylistGroup, PlaylistItem} from "../../model/playlist";
+import './playlist-gallery.scss';
+import React, {useCallback, useEffect, useState} from "react";
+import {useSnackbar} from "notistack";
+import {getIconByName} from "../../icons/icons";
 import copyToClipboard from "../../utils/clipboard";
 import {first} from "rxjs/operators";
 import {noop} from "rxjs";
-import {useSnackbar} from "notistack";
-import {getIconByName} from "../../icons/icons";
-import ServerConfig from "../../model/server-config";
 
-export type PlaylistTreeState = { [key: number]: boolean };
-
-interface PlaylistTreeProps {
+interface PlaylistGalleryProps {
     serverConfig: ServerConfig;
     data: PlaylistGroup[];
-    state: PlaylistTreeState;
     onCopy: (playlistItem: PlaylistItem) => void;
     onPlay?: (playlistItem: PlaylistItem) => void;
     onDownload?: (playlistItem: PlaylistItem) => void;
     onWebSearch?: (playlistItem: PlaylistItem) => void;
+
 }
 
-export default function PlaylistTree(props: PlaylistTreeProps) {
-    const {serverConfig, state, data, onCopy, onPlay, onDownload, onWebSearch} = props;
+export default function PlaylistGallery(props: PlaylistGalleryProps) {
 
-    const [, setForceUpdate] = useState(null);
-    const expanded = useRef<PlaylistTreeState>({});
+    const {data, onCopy, onPlay, onWebSearch, onDownload, serverConfig} = props;
     const {enqueueSnackbar/*, closeSnackbar*/} = useSnackbar();
-    const [videoExtensions, setVideoExtensions] = useState<string[]>([]);
+
+    const [showcaseGroup, setShowcaseGroup] = useState<PlaylistGroup>();
 
     useEffect(() => {
-        if (serverConfig) {
-            setVideoExtensions(serverConfig.video?.extensions);
-        }
-        return noop;
-    }, [serverConfig]);
+        setShowcaseGroup(data.length ? data[0] : undefined)
+    }, [data]);
 
     const getPlaylistItemById = useCallback((itemId: string): PlaylistItem => {
         const id = parseInt(itemId);
@@ -50,18 +44,6 @@ export default function PlaylistTree(props: PlaylistTreeProps) {
         }
         return undefined;
     }, [data]);
-
-    const handleChange = useCallback((event: any) => {
-        const key = event.target.dataset.group;
-        state[key] = !state[key];
-        setForceUpdate({});
-    }, [state]);
-
-    const handleExpand = useCallback((event: any) => {
-        const key = event.target.dataset.group;
-        expanded.current[key] = !expanded.current[key];
-        setForceUpdate({});
-    }, []);
 
     const handleClipboardUrl = useCallback((e: any) => {
         const item = getPlaylistItemById(e.target.dataset.item);
@@ -82,7 +64,7 @@ export default function PlaylistTree(props: PlaylistTreeProps) {
                 onWebSearch(item);
             }
         }
-   }, [getPlaylistItemById, onWebSearch]);
+    }, [getPlaylistItemById, onWebSearch]);
 
     const handleDownloadUrl = useCallback((e: any) => {
         if (onDownload) {
@@ -107,71 +89,80 @@ export default function PlaylistTree(props: PlaylistTreeProps) {
     }, [onPlay, getPlaylistItemById]);
 
     const isVideoFile = useCallback((entry: PlaylistItem): boolean => {
-        if (videoExtensions && entry.header.url) {
-            for (const ext of videoExtensions) {
+        if (serverConfig?.video?.extensions && entry.header.url) {
+            for (const ext of serverConfig.video.extensions) {
                 if (entry.header.url.endsWith(ext)) {
                     return true;
                 }
             }
         }
         return false;
-    }, [videoExtensions]);
+    }, [serverConfig]);
 
-    const renderEntry = useCallback((entry: PlaylistItem, index: number): React.ReactNode => {
-        return <div key={entry.id} className={'tree-channel'}>
-            <div className={'tree-channel-tools'}>
-                <div className={'tool-button'} data-item={entry.id} onClick={handleClipboardUrl}>
+    const handleGroupClick = useCallback((event: any) => {
+        const key = parseInt(event.target.dataset.group);
+        const group = data.find(plg => plg.id === key);
+        setShowcaseGroup(group);
+    }, [data]);
+
+    const renderGroup = useCallback((group: PlaylistGroup): React.ReactNode => {
+        return <div className={'playlist-gallery__categories-group' + (showcaseGroup?.id === group.id ? ' selected-group' : '') } key={group.id} data-group={group.id}
+                    onClick={handleGroupClick}>
+            {group.title}
+        </div>
+    }, [handleGroupClick, showcaseGroup]);
+
+    const renderPlaylistItem = useCallback((playlistItem: PlaylistItem) => {
+        return <div className="playlist-gallery__showcase-card" key={playlistItem.id}>
+            <div className="playlist-gallery__showcase-card-header">
+                {playlistItem.header.name}
+            </div>
+            <div className="playlist-gallery__showcase-card-content">
+                <img  alt="logo" src={playlistItem.header.logo || playlistItem.header.logo_small || 'placeholder.png'}/>
+            </div>
+            <div className="playlist-gallery__showcase-card-footer">
+                <div className={'tool-button'} data-item={playlistItem.id} onClick={handleClipboardUrl}>
                     {getIconByName('LinkRounded')}
                 </div>
-                <div style={{display: 'none'}} className={'tool-button'} data-item={entry.id} onClick={handlePlayUrl}>
+                <div style={{display: 'none'}} className={'tool-button'} data-item={playlistItem.id}
+                     onClick={handlePlayUrl}>
                     {getIconByName('PlayArrow')}
                 </div>
-                {isVideoFile(entry) &&
+                {isVideoFile(playlistItem) &&
                     <>
-                        <div className={'tool-button'} data-item={entry.id} onClick={handleDownloadUrl}>
+                        <div className={'tool-button'} data-item={playlistItem.id} onClick={handleDownloadUrl}>
                             {getIconByName('Download')}
                         </div>
                         {serverConfig.video?.web_search &&
-                            <div className={'tool-button'} data-item={entry.id} onClick={handleWebSearch}>
+                            <div className={'tool-button'} data-item={playlistItem.id} onClick={handleWebSearch}>
                                 {getIconByName('WebSearch')}
                             </div>
                         }
                     </>
                 }
             </div>
-            <div className={'tree-channel-content'}>
-                <div className={'tree-channel-nr'}>{index + 1}</div>
-                {entry.header.name}</div>
         </div>
-    }, [handleClipboardUrl, handlePlayUrl, handleDownloadUrl, isVideoFile, handleWebSearch, serverConfig]);
+    }, [handleClipboardUrl, handleDownloadUrl, handlePlayUrl, handleWebSearch, isVideoFile, serverConfig]);
 
-    const renderGroup = useCallback((group: PlaylistGroup): React.ReactNode => {
-        return <div className={'tree-group'} key={group.id}>
-            <div className={'tree-group-header'}>
-                <div className={'tree-expander'} data-group={group.id}
-                     onClick={handleExpand}>{getIconByName(expanded.current[group.id] ?
-                    'ExpandMore' : 'ChevronRight')}</div>
-                <div className={'tree-group-header-content'}>
-                    <input type={"checkbox"} onChange={handleChange} data-group={group.id}/>
-                    {group.title}
-                    <div className={'tree-group-count'}>({group.channels.length})</div>
-                </div>
-            </div>
-            {expanded.current[group.id] && (
-                <div className={'tree-group-childs'}>
-                    {group.channels.map(renderEntry)}
-                </div>)}
-        </div>;
-    }, [handleChange, handleExpand, renderEntry]);
+    const renderShowcase = useCallback(() => {
+        if (!showcaseGroup) {
+            return <React.Fragment></React.Fragment>
+        }
+        return showcaseGroup.channels.map(renderPlaylistItem);
+    }, [showcaseGroup, renderPlaylistItem]);
 
-    const renderPlaylist = useCallback((): React.ReactNode => {
-        if (!data) {
+
+    const renderCategories = useCallback(() => {
+        if (!data?.length) {
             return <React.Fragment/>;
         }
-        return <React.Fragment>
-            {data.map(renderGroup)}
-        </React.Fragment>;
+        return <div className="playlist-gallery__categories">{data.map(renderGroup)}</div>
     }, [data, renderGroup]);
 
-    return <div className={'playlist-tree'}>{renderPlaylist()}</div>;
-} 
+    return <div className="playlist-gallery">
+        {renderCategories()}
+        <div className="playlist-gallery__showcase">
+            {renderShowcase()}
+        </div>
+    </div>
+}