+++ src/model/config.rs
@@ -574,6 +574,8 @@ pub(crate) struct Config {
     #[serde(skip_serializing, skip_deserializing)]
     pub _api_proxy: Arc<RwLock<Option<ApiProxyConfig>>>,
     #[serde(skip_serializing, skip_deserializing)]
+    pub _config_path: String,
+    #[serde(skip_serializing, skip_deserializing)]
     pub _config_file_path: String,
     #[serde(skip_serializing, skip_deserializing)]
     pub _sources_file_path: String,