+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
-# v1.1.8(2024-02-??)
+# v1.1.8(2024-03-??)
 * Fixed WebUI Option-Select  
-* WebUI: added gallery view as second view for playlist 
+* WebUI: added gallery view as second view for playlist
+* Breaking change config path. The config path is now default ./config. 
+  You can provide a config path with the "-p" argument.
 
 # v1.1.7(2024-01-30)
 * Renamed api-proxy.yml server info field `ip` to `host`